backend/requirements.txt
mysql-connector-python~=8.4.0
bcrypt~=4.1.3
flask~=3.0.3
flask-cors~=4.0.1
python-dotenv~=1.0.1
flask-jwt-extended~=4.6.0
